+++ src/split/esb/common.h
@@ -107,7 +107,7 @@ uint32_t get_nonce();
 int process_payload(char* data, size_t length, uint32_t nonce);
 
 int make_packet(struct esb_payload *payload);
-size_t handle_packet();
+int handle_packet();
 void reset_buffers();
 
 int tx_msgq_init(int *type_to_idx);
